@@ -5,11 +5,13 @@
 
 	<properties>
 		<java.version>17</java.version>
-		<spring-boot.version>3.0.5</spring-boot.version>
+		<spring-boot.version>3.1.0</spring-boot.version>
 		<json-web-token.version>0.11.5</json-web-token.version>
 		<lombok.version>1.18.22</lombok.version>
-		<liquibase.version>4.17.1</liquibase.version>
-		<auth0-jwt.version>4.2.1</auth0-jwt.version>
+		<liquibase.version>4.22.0</liquibase.version>
+		<auth0-jwt.version>4.3.0</auth0-jwt.version>
+		<mapstruct.version>1.5.5.Final</mapstruct.version>
+		<hibernate.version>6.2.3.Final</hibernate.version>
 	</properties>
 
 	<parent>
@@ -23,6 +25,12 @@
 	<version>0.0.1-SNAPSHOT</version>
 	<name>basketstats</name>
 	<description>Basketstats system</description>
+
+<!--	<profiles>-->
+<!--		<profile>-->
+<!--			<id>profile-1</id>-->
+<!--		</profile>-->
+<!--	</profiles>-->
 
 	<dependencies>
 		<!-- SPRING  -->
@@ -58,7 +66,7 @@
 		<dependency>
 			<groupId>com.google.code.gson</groupId>
 			<artifactId>gson</artifactId>
-			<version>2.9.0</version>
+			<version>2.10.1</version>
 		</dependency>
 
 		<!-- Auth0 -->
@@ -73,11 +81,18 @@
 			<version>${auth0-jwt.version}</version>
 		</dependency>
 
-		<!-- DB  -->
+		<!-- DB - MS SQL SERVER -->
 		<dependency>
 			<groupId>com.microsoft.sqlserver</groupId>
 			<artifactId>mssql-jdbc</artifactId>
 			<scope>runtime</scope>
+		</dependency>
+
+		<!-- DB - POSTGRES -->
+		<dependency>
+			<groupId>org.postgresql</groupId>
+			<artifactId>postgresql</artifactId>
+			<version>42.6.0</version>
 		</dependency>
 
 		<!-- LIQUIBASE -->
@@ -98,29 +113,33 @@
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
+			<scope>test</scope>
 		</dependency>
-
 		<dependency>
 			<groupId>org.springframework</groupId>
 			<artifactId>spring-test</artifactId>
+			<scope>test</scope>
 		</dependency>
-
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-test-autoconfigure</artifactId>
+			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-test</artifactId>
 			<version>${spring-boot.version}</version>
+			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.mockito</groupId>
 			<artifactId>mockito-core</artifactId>
+			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.assertj</groupId>
 			<artifactId>assertj-core</artifactId>
+			<scope>test</scope>
 		</dependency>
 
 		<dependency>
@@ -147,6 +166,30 @@
 			<version>3.0.1</version>
 		</dependency>
 
+		<dependency>
+			<groupId>org.hibernate.orm</groupId>
+			<artifactId>hibernate-core</artifactId>
+			<version>${hibernate.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.mapstruct</groupId>
+			<artifactId>mapstruct</artifactId>
+			<version>${mapstruct.version}</version>
+		</dependency>
+
+		<!-- FTP Server -->
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-integration</artifactId>
+			<version>3.1.5</version>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.integration</groupId>
+			<artifactId>spring-integration-ftp</artifactId>
+			<version>6.1.4</version>
+		</dependency>
+
 	</dependencies>
 
 	<build>
@@ -171,7 +214,7 @@
 						<path>
 							<groupId>org.mapstruct</groupId>
 							<artifactId>mapstruct-processor</artifactId>
-							<version>1.5.2.Final</version>
+							<version>${mapstruct.version}</version>
 						</path>
 						<path>
 							<groupId>org.projectlombok</groupId>
