@@ -10,6 +10,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import reforged.marcin.krysiak.basketstats.dto.LeagueDto;
 import reforged.marcin.krysiak.basketstats.exceptions.LeagueNotFoundException;
 import reforged.marcin.krysiak.basketstats.models.League;
 import reforged.marcin.krysiak.basketstats.service.LeagueService;
@@ -25,7 +26,7 @@ public class LeagueController {
 
     @GetMapping
     public ResponseEntity<Page<League>> getLeaguesBySpec(
-            @Spec(path = "name", params = "nameLike", spec = LikeIgnoreCase.class)
+            @Spec(path = "name", params = "searchLike", spec = LikeIgnoreCase.class)
             Specification<League> spec,
             Pageable pageable
     ) {
@@ -33,8 +34,11 @@ public class LeagueController {
     }
 
     @GetMapping("/get-all")
-    public ResponseEntity<List<League>> getAllLeagues() {
-        return ResponseEntity.ok(leagueService.getAll());
+    public ResponseEntity<List<League>> getAllLeagues(
+            @Spec(path = "name", params = "searchLike", spec = LikeIgnoreCase.class)
+            Specification<League> spec
+    ) {
+        return ResponseEntity.ok(leagueService.getAll(spec));
     }
 
     @GetMapping("/{id}")
@@ -56,15 +60,14 @@ public class LeagueController {
     }
 
     @PostMapping
-    public ResponseEntity<League> createLeague(@Valid @RequestBody League league) {
-        return ResponseEntity.ok().body(leagueService.create(league));
+    public ResponseEntity<LeagueDto> createLeague(@Valid @RequestBody LeagueDto league) {
+        return ResponseEntity.ok(leagueService.create(league));
     }
 
     @PutMapping("/{id}")
-    public Object updateUser(@Valid @RequestBody League league, @PathVariable Long id) {
+    public ResponseEntity<LeagueDto> updateUser(@Valid @RequestBody LeagueDto league, @PathVariable Long id) {
 
-        leagueService.update(id, league);
-        return true;
+        return ResponseEntity.ok(leagueService.update(id, league));
     }
 
     @DeleteMapping("/{id}")
