@@ -1,21 +1,27 @@
 package reforged.marcin.krysiak.basketstats.controllers;
 
+import net.kaczmarzyk.spring.data.jpa.domain.Equal;
+import net.kaczmarzyk.spring.data.jpa.domain.LikeIgnoreCase;
+import net.kaczmarzyk.spring.data.jpa.web.annotation.And;
+import net.kaczmarzyk.spring.data.jpa.web.annotation.Spec;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import reforged.marcin.krysiak.basketstats.dto.PlayerStatsAvgForTeamDTO;
-import reforged.marcin.krysiak.basketstats.dto.TeamWithLogoDTO;
-import reforged.marcin.krysiak.basketstats.exceptions.ApiError;
+import reforged.marcin.krysiak.basketstats.dto.TeamDto;
+import reforged.marcin.krysiak.basketstats.dto.TeamFormDTO;
 import reforged.marcin.krysiak.basketstats.exceptions.TeamNotFoundException;
-import reforged.marcin.krysiak.basketstats.messages.ResponseMessage;
 import reforged.marcin.krysiak.basketstats.models.Team;
 import reforged.marcin.krysiak.basketstats.service.PlayerStatsService;
 import reforged.marcin.krysiak.basketstats.service.TeamService;
+import reforged.marcin.krysiak.basketstats.service.ftp.FtpService;
 
 import java.util.List;
+import java.util.Objects;
 
 @RequestMapping("/api/team")
 @RestController
@@ -28,94 +34,81 @@ public class TeamController {
     @Autowired
     PlayerStatsService playerStatsService;
 
+    @Autowired
+    FtpService ftpService;
 
-    @GetMapping("/list")
-    public ResponseEntity<List<Team>> getAllTeams() {
-
-        return ResponseEntity.ok().body(teamService.getAllTeams());
+    @GetMapping
+    public ResponseEntity<Page<TeamDto>> getTeamsBySpec(
+            @And({
+                    @Spec(path = "name", params = "searchLike", spec = LikeIgnoreCase.class),
+                    @Spec(path = "league.id", params = "leagueIdEqual", spec = Equal.class)
+            })
+            Specification<Team> spec,
+            Pageable pageable
+    ) {
+        return ResponseEntity.ok(teamService.getAllBySpec(spec, pageable));
     }
 
-    @GetMapping("/leagueList/{id}")
-    public ResponseEntity<List<Team>> getAllTeamsByLeague(@PathVariable Long id) {
+    @GetMapping("/league-id/{leagueId}")
+    public ResponseEntity<Page<TeamDto>> getTeamsForLeagueBySpec(
+            @PathVariable Long leagueId,
+            @And({
+                    @Spec(path = "name", params = "searchLike", spec = LikeIgnoreCase.class),
+            })
+            Specification<Team> spec,
+            Pageable pageable
+    ) {
+        return ResponseEntity.ok(teamService.getAllForLeagueBySpec(leagueId, spec, pageable));
+    }
 
-        return ResponseEntity.ok().body(teamService.getAllTeamsByLeague(id));
+    @GetMapping("/get-all")
+    public ResponseEntity<List<TeamDto>> getAllTeams(
+            @And({
+                    @Spec(path = "name", params = "searchLike", spec = LikeIgnoreCase.class),
+                    @Spec(path = "league.id", params = "leagueIdEqual", spec = Equal.class)
+            })
+            Specification<Team> spec
+    ) {
+        return ResponseEntity.ok(teamService.getAll(spec));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Team> getTeamById(@PathVariable Long id) throws Exception {
-        Team team = teamService.getTeamById(id)
-                .orElseThrow(TeamNotFoundException::new);
-
-        return ResponseEntity.ok().body(team);
+    public ResponseEntity<TeamDto> getTeamById(@PathVariable Long id) throws Exception {
+        return ResponseEntity.ok(teamService.getById(id));
     }
 
     @GetMapping("/playersAvgStats/{id}")
     public ResponseEntity<List<PlayerStatsAvgForTeamDTO>> getAllAvgPlayerStatsForTeam(@PathVariable Long id) throws Exception {
 
-        return ResponseEntity.ok().body(playerStatsService.getAllAvgPlayerStatsForTeam(id));
+        return ResponseEntity.ok(playerStatsService.getAllAvgPlayerStatsForTeam(id));
     }
 
-    @PostMapping("/create")
-    public ResponseEntity<ResponseMessage> createTeam(@ModelAttribute("team") TeamWithLogoDTO team) {
-
-        String message = "";
-        try {
-            teamService.createTeam(team);
-
-            message = "Drużyna została dodana pomyślnie wraz z obrazem: " + team.getLogoFile().getName();
-            return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
-        } catch (Exception e) {
-            message = "Drużyna nie została zaktualizowana, problem z obrazem: " + team.getLogoFile().getName() + "!";
-            return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponseMessage(message));
-        }
-    }
-
-    @GetMapping("/logo/{id}")
-    public ResponseEntity<byte[]> getFile(@PathVariable Long id) {
-        Team team = teamService.getTeamById(id)
-                .orElseGet(Team::new);
-
-        return ResponseEntity.ok()
-                .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + team.getLogo() + "\"")
-                .body(team.getData());
+    @PostMapping
+    public ResponseEntity<TeamDto> create(@ModelAttribute("team") TeamFormDTO team) {
+        return ResponseEntity.ok(teamService.create(team));
     }
 
     @GetMapping("/isTeamWithNameExisting/{name}")
     public boolean isUserWithEmailExisting(@PathVariable String name) throws TeamNotFoundException {
-        try {
-            Team team = teamService.getTeamByName(name);
-            if(team == null) {
-                throw new TeamNotFoundException();
-            }
-        } catch (TeamNotFoundException e) {
-            return false;
-        }
-        return true;
+        return Objects.nonNull(teamService.getByName(name));
     }
 
-    @PutMapping("/update/{id}")
-    public ResponseEntity<ResponseMessage> updateUser(@ModelAttribute("team") TeamWithLogoDTO team, @PathVariable Long id) {
+    @PutMapping("/{id}")
+    public ResponseEntity<TeamDto> update(@ModelAttribute("team") TeamFormDTO team, @PathVariable Long id) {
 
-        String message = "";
-        try {
-            teamService.updateTeam(id, team);
-
-            message = "Drużyna została zaktualizowana pomyślnie wraz z obrazem: " + team.getLogoFile().getName();
-            return ResponseEntity.status(HttpStatus.OK).body(new ResponseMessage(message));
-        } catch (Exception e) {
-            message = "Drużyna nie została zaktualizowana, problem z obrazem: " + team.getLogoFile().getName() + "!";
-            return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponseMessage(message));
-        }
+        return ResponseEntity.ok(teamService.update(id, team));
     }
 
-    @DeleteMapping("/delete/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteTeam(@PathVariable Long id) {
 
         try {
-            teamService.deleteTeam(id);
+            teamService.delete(id);
             return ResponseEntity.noContent().build();
         } catch (EmptyResultDataAccessException e) {
             return ResponseEntity.notFound().build();
         }
     }
+
+
 }
