@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.Setter;
+import reforged.marcin.krysiak.basketstats.enums.PlayerPosition;
 
 @Getter
 @Setter
@@ -13,7 +14,7 @@ public class Player {
 
     @Id
     @SequenceGenerator(allocationSize = 1, name = "players_id_seq", sequenceName = "players_id_seq")
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "players_id_seq")
     @Column(name = "id")
     private Long id;
 
@@ -28,12 +29,12 @@ public class Player {
     @Column(name = "number")
     private int number;
 
-    @ManyToOne(cascade = CascadeType.DETACH)
-    @JoinColumn(name = "team_id", referencedColumnName = "id", nullable = false)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "team_id", nullable = false)
     private Team team;
 
-    @Column(name = "position", length = 50)
-    private String position;
+    @Enumerated(EnumType.STRING)
+    private PlayerPosition position;
 
     @Column(name = "captain")
     private boolean captain;
