@@ -1,69 +1,35 @@
 package reforged.marcin.krysiak.basketstats.models;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.Arrays;
+import reforged.marcin.krysiak.basketstats.models.base.AuditEntity;
 
 @Getter
 @Setter
 @Entity
 @Table(name = "teams")
-public class Team {
+@NoArgsConstructor
+@AllArgsConstructor
+public class Team extends AuditEntity {
 
     @Id
     @SequenceGenerator(allocationSize = 1, name = "teams_id_seq", sequenceName = "teams_id_seq")
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "teams_id_seq")
     private Long id;
 
     @Column(length = 50, nullable = false)
     private String name;
 
-    @Column(nullable = true)
-    private String logo;
+    private String fileName;
 
-    @Column(nullable = true)
-    @Basic(fetch = FetchType.LAZY)
-    private byte[] data;
+    private String filePath;
 
-    @Column(length = 100, nullable = true)
-    private String type;
+    private String fileType;
 
-    @ManyToOne()
-    @JoinColumn(name = "league_id", referencedColumnName = "id", nullable = false)
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "league_id", nullable = false)
     private League league;
-
-    public Team() {
-
-    }
-
-    public Team(Long id, String name, String logo, League league) {
-        this.id = id;
-        this.name = name;
-        this.logo = logo;
-        this.league = league;
-    }
-
-    public Team(Long id, String name, String logo, byte[] data, String type, League league) {
-        this.id = id;
-        this.name = name;
-        this.logo = logo;
-        this.data = data;
-        this.type = type;
-        this.league = league;
-    }
-
-    @Override
-    public String toString() {
-        return "Team{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", logo='" + logo + '\'' +
-                ", data=" + Arrays.toString(data) +
-                ", type='" + type + '\'' +
-                ", league=" + league +
-                '}';
-    }
 }
