@@ -1,5 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { delay, first } from "rxjs/operators";
+import { select, Store } from '@ngrx/store';
+import { Observable } from 'rxjs';
+import { delay } from "rxjs/operators";
+import * as LeagueActions from 'src/app/components/league/state/league.actions';
+import * as LeagueSelector from 'src/app/components/league/state/league.selectors';
+import { BasketStatsAppState } from 'src/app/store';
 import { League } from "../../models/league.model";
 import { Match } from "../../models/match";
 import { LeagueService } from "../league/league.service";
@@ -17,7 +22,10 @@ export class ScheduleComponent implements OnInit {
 
   isLoading = false;
 
+  public leagues$: Observable<League[]>;
+
   constructor(
+    private store: Store<BasketStatsAppState>,
     private leagueService: LeagueService,
     private matchService: MatchService
   ) { }
@@ -27,15 +35,21 @@ export class ScheduleComponent implements OnInit {
   }
 
   loadData() {
-    this.isLoading = true;
-    this.leagueService.getAll().pipe(first()).subscribe(data => {
-      this.leagueList = data;
-    });
+    this.loadLeagues();
+    this.leagues$ = this.store.pipe(select(LeagueSelector.selectAllLeagues));
+    // this.isLoading = true;
+    // this.leagueService.getAll().pipe(first()).subscribe(data => {
+    //   this.leagueList = data;
+    // });
 
     this.matchService.getMatchList().pipe(delay(500)).subscribe(data => {
       this.matchList = data;
       this.isLoading = false;
     });
+  }
+
+  private loadLeagues(): void {
+    this.store.dispatch(LeagueActions.loadAllLeagues());
   }
 
   getMatchesOfLeague(leagueId: number) {
