@@ -1,9 +1,9 @@
+import { FileDetails } from "../shared/model/image-file";
 import { League } from "./league.model";
 
 export class Team {
   id: number;
   name: string;
-  logo: string;
   league: League;
-  logoFile: Blob;
+  file: FileDetails;
 }
