@@ -3,15 +3,29 @@ import { TableOperation } from "../enum/table-operation";
 
 export interface TableConfig {
     tableColumns: TableColumn[];
-    headerText: string;
+    headerText?: string;
     searchPlaceholder: string;
     addButtonTooltip: string;
     pageSizeOptions?: number[];
     tableOperations?: TableOperation[];
+    previewRoute?: string;
+    editRoute?: string;
 }
 
 export interface TableColumn {
     name: string;
     type: ColumnType;
     translation?: string;
+}
+
+export function getCopyWithColumnIfNotExists(tableColumns: TableColumn[], column: TableColumn, first?: boolean) {
+    tableColumns = !!tableColumns ? tableColumns?.filter(tableColumn => tableColumn.name !== column.name) : [];
+
+    if (first) {
+        tableColumns.unshift(column);
+    } else {
+        tableColumns.push(column);
+    }
+
+    return tableColumns;
 }
